feat: Implementing users and lawyers
A base user class (User) was implemented to serve as a generic model for future user types within the CRM system. Based on this structure, the Lawyer model was created, inheriting from the User class and establishing an inheritance relationship with distinct tables in the database. In this context, a lawyer is considered a type of client, although not every client is necessarily a system user.

Login and logout routes were also implemented, returning a session token with a default validity of 30 days (this value may be adjusted in the future). Every interaction with the application requires the user to be authenticated. Therefore, a validation function was created to verify the session token and ensure that the request is coming from the same IP address as the original session.

In addition, the following routes were developed specifically for the Lawyer model:

- POST / Create: registers a new lawyer;
- PUT / Update: updates the details of an existing lawyer;
- GET / All: returns all lawyers registered in the system;
- GET / One: returns a specific lawyer based on their id;
- DELETE / One: deletes a specific lawyer based on their id;
- DELETE / Multiple: allows multiple lawyers to be deleted by sending a list of ids;
- PUT / Activate or Deactivate Multiple: updates the is_active field for multiple users simultaneously;
- POST / Upload Profile Picture: allows an image to be uploaded and associated as the lawyer's profile picture.